Count explicit negatives in text.
The regular expression in line 53 is what was used to generate the data in the MEDINFO 2017 paper.  It has a bug that affected the counts in CRAFT, but not in the clinical data.  The regular expression in line 54 fixes that bug by adding a global switch to the regular expression.  That's what was used to generate the corresponding data in the LREC 2018 paper.